+++ Microsoft.Toolkit.HighPerformance/Extensions/MemoryExtensions.cs
@@ -8,6 +8,8 @@ using System.Diagnostics.Contracts;
 using System.IO;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
+using Microsoft.Toolkit.HighPerformance.Buffers.Internals;
+using Microsoft.Toolkit.HighPerformance.Buffers.Internals.Interfaces;
 using MemoryStream = Microsoft.Toolkit.HighPerformance.Streams.MemoryStream;
 
 namespace Microsoft.Toolkit.HighPerformance.Extensions
@@ -83,206 +85,6 @@ namespace Microsoft.Toolkit.HighPerformance.Extensions
         public static Stream AsStream(this Memory<byte> memory)
         {
             return new MemoryStream(memory);
-        }
-
-        /// <summary>
-        /// An interface for a <see cref="MemoryManager{T}"/> instance that can reinterpret its underlying data.
-        /// </summary>
-        internal interface IMemoryManager
-        {
-            /// <summary>
-            /// Creates a new <see cref="MemoryManager{T}"/> that reinterprets the underlying data for the current instance.
-            /// </summary>
-            /// <typeparam name="T">The target type to cast the items to.</typeparam>
-            /// <param name="offset">The starting offset within the data store.</param>
-            /// <param name="length">The original used length for the data store.</param>
-            /// <returns>A new <see cref="MemoryManager{T}"/> instance of the specified type, reinterpreting the current items.</returns>
-            MemoryManager<T> Cast<T>(int offset, int length)
-                where T : unmanaged;
-        }
-
-        /// <summary>
-        /// A custom <see cref="MemoryManager{T}"/> that casts data from a <typeparamref name="TFrom"/> array, to <typeparamref name="TTo"/> values.
-        /// </summary>
-        /// <typeparam name="TFrom">The source type of items to read.</typeparam>
-        /// <typeparam name="TTo">The target type to cast the source items to.</typeparam>
-        internal sealed class ArrayMemoryManager<TFrom, TTo> : MemoryManager<TTo>, IMemoryManager
-            where TFrom : unmanaged
-            where TTo : unmanaged
-        {
-            /// <summary>
-            /// The source <typeparamref name="TFrom"/> array to read data from.
-            /// </summary>
-            private readonly TFrom[] array;
-
-            /// <summary>
-            /// The starting offset within <see name="array"/>.
-            /// </summary>
-            private readonly int offset;
-
-            /// <summary>
-            /// The original used length for <see name="array"/>.
-            /// </summary>
-            private readonly int length;
-
-            /// <summary>
-            /// Initializes a new instance of the <see cref="ArrayMemoryManager{TFrom, TTo}"/> class.
-            /// </summary>
-            /// <param name="array">The source <typeparamref name="TFrom"/> array to read data from.</param>
-            /// <param name="offset">The starting offset within <paramref name="array"/>.</param>
-            /// <param name="length">The original used length for <paramref name="array"/>.</param>
-            public ArrayMemoryManager(TFrom[] array, int offset, int length)
-            {
-                this.array = array;
-                this.offset = offset;
-                this.length = length;
-            }
-
-            /// <inheritdoc/>
-            public override Span<TTo> GetSpan()
-            {
-#if SPAN_RUNTIME_SUPPORT
-                ref TFrom r0 = ref this.array.DangerousGetReferenceAt(this.offset);
-
-                Span<TFrom> span = MemoryMarshal.CreateSpan(ref r0, this.length);
-#else
-                Span<TFrom> span = this.array.AsSpan(this.offset, this.length);
-#endif
-
-                // We rely on MemoryMarshal.Cast here to deal with calculating the effective
-                // size of the new span to return. This will also make the behavior consistent
-                // for users that are both using this type as well as casting spans directly.
-                return MemoryMarshal.Cast<TFrom, TTo>(span);
-            }
-
-            /// <inheritdoc/>
-            public override unsafe MemoryHandle Pin(int elementIndex = 0)
-            {
-                int
-                    bytePrefix = this.offset + Unsafe.SizeOf<TFrom>(),
-                    byteSuffix = elementIndex * Unsafe.SizeOf<TTo>(),
-                    byteOffset = bytePrefix + byteSuffix;
-
-                GCHandle handle = GCHandle.Alloc(this.array, GCHandleType.Pinned);
-
-                ref TFrom r0 = ref this.array.DangerousGetReference();
-                ref byte r1 = ref Unsafe.As<TFrom, byte>(ref r0);
-                ref byte r2 = ref Unsafe.Add(ref r1, byteOffset);
-                void* pi = Unsafe.AsPointer(ref r2);
-
-                return new MemoryHandle(pi, handle);
-            }
-
-            /// <inheritdoc/>
-            public override void Unpin()
-            {
-            }
-
-            /// <inheritdoc/>
-            protected override void Dispose(bool disposing)
-            {
-            }
-
-            /// <inheritdoc/>
-            public MemoryManager<T> Cast<T>(int offset, int length)
-                where T : unmanaged
-            {
-                return new ArrayMemoryManager<TFrom, T>(this.array, this.offset + offset, length);
-            }
-        }
-
-        /// <summary>
-        /// A custom <see cref="MemoryManager{T}"/> that casts data from a <see cref="MemoryManager{T}"/> of <typeparamref name="TFrom"/>, to <typeparamref name="TTo"/> values.
-        /// </summary>
-        /// <typeparam name="TFrom">The source type of items to read.</typeparam>
-        /// <typeparam name="TTo">The target type to cast the source items to.</typeparam>
-        internal sealed class ProxyMemoryManager<TFrom, TTo> : MemoryManager<TTo>, IMemoryManager
-            where TFrom : unmanaged
-            where TTo : unmanaged
-        {
-            /// <summary>
-            /// The source <see cref="MemoryManager{T}"/> to read data from.
-            /// </summary>
-            private readonly MemoryManager<TFrom> memoryManager;
-
-            /// <summary>
-            /// The starting offset within <see name="memoryManager"/>.
-            /// </summary>
-            private readonly int offset;
-
-            /// <summary>
-            /// The original used length for <see name="memoryManager"/>.
-            /// </summary>
-            private readonly int length;
-
-            /// <summary>
-            /// Initializes a new instance of the <see cref="ProxyMemoryManager{TFrom, TTo}"/> class.
-            /// </summary>
-            /// <param name="memoryManager">The source <see cref="MemoryManager{T}"/> to read data from.</param>
-            /// <param name="offset">The starting offset within <paramref name="memoryManager"/>.</param>
-            /// <param name="length">The original used length for <paramref name="memoryManager"/>.</param>
-            public ProxyMemoryManager(MemoryManager<TFrom> memoryManager, int offset, int length)
-            {
-                this.memoryManager = memoryManager;
-                this.offset = offset;
-                this.length = length;
-            }
-
-            /// <inheritdoc/>
-            public override Span<TTo> GetSpan()
-            {
-                Span<TFrom> span = this.memoryManager.GetSpan().Slice(this.offset, this.length);
-
-                return MemoryMarshal.Cast<TFrom, TTo>(span);
-            }
-
-            /// <inheritdoc/>
-            public override MemoryHandle Pin(int elementIndex = 0)
-            {
-                int byteOffset = elementIndex * Unsafe.SizeOf<TTo>();
-
-#if NETSTANDARD1_4
-                int
-                    shiftedOffset = byteOffset / Unsafe.SizeOf<TFrom>(),
-                    remainder = byteOffset - (shiftedOffset * Unsafe.SizeOf<TFrom>());
-#else
-                int shiftedOffset = Math.DivRem(byteOffset, Unsafe.SizeOf<TFrom>(), out int remainder);
-#endif
-
-                if (remainder != 0)
-                {
-                    ThrowArgumentExceptionForInvalidAlignment();
-                }
-
-                return this.memoryManager.Pin(this.length + shiftedOffset);
-            }
-
-            /// <inheritdoc/>
-            public override void Unpin()
-            {
-                this.memoryManager.Unpin();
-            }
-
-            /// <inheritdoc/>
-            protected override void Dispose(bool disposing)
-            {
-                ((IDisposable)this.memoryManager).Dispose();
-            }
-
-            /// <inheritdoc/>
-            public MemoryManager<T> Cast<T>(int offset, int length)
-                where T : unmanaged
-            {
-                return new ProxyMemoryManager<TFrom, T>(this.memoryManager, this.offset + offset, length);
-            }
-
-            /// <summary>
-            /// Throws an <see cref="ArgumentOutOfRangeException"/> when <see cref="Pin"/> receives an invalid target index.
-            /// </summary>
-            private static void ThrowArgumentExceptionForInvalidAlignment()
-            {
-                throw new ArgumentOutOfRangeException("elementIndex", "The input index doesn't result in an aligned item access");
-            }
         }
     }
 }
